Enabling optional Azure MI based AUTH
